+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * (theimo1221) Preparations for the implementation of SmartGardenDevices
 * (theimo1221) Add Smart-Garden Sensor (temperature and humidity of soil)
 * (theimo1221) Add BatteryDriven to Smart Garden base device
+* (theimo1221) Add Smartgarden mover as an actuator until we have iMover
 
 ## 3.0.0-alpha.61 (2024-05-10)